Add Qodana code quality tool
This commit adds the Qodana, a static code analysis tool, to the CI/CD pipeline and does a clean-up in config.go. This is done for maintaining code quality and standards. A new Qodana configuration file(qodana.yaml) has been added and the Qodana GitHub action has been configured in a new workflow file(qodana_code_quality.yml) to run on push and pull_request events on master branch. A 'getEnvBool' function was removed from config.go file as it was not being used anywhere else in the project.