Refactor MongoDB connection and server initialization logic
This commit refactors the MongoDB connection and server initialization logic into separate functions for better modularization and readability.
Previously, this logic was directly in the main function, making it difficult to manage. Now, NewMongoDB and NewServer functions have been created in server.go, improving the organization of the code. Moreover, new catalog-service and catalog-handler files have been added under the `src` directory to separate the handling of catalog APIs.
Also, the port number and logger level values have been updated in the .env file. These changes together enhance the readability and maintainability of the code.