Refactor code and update swagger docs
Refactored error and response handling structure from the `models` microservice to `handlers`. The port number in the `.env` file was also changed from 8089 to 8080. Updates to the `swagger.json` and `docs.go` files reflect these structural changes, along with an enhancement of the description. The route name in the handler function was changed from `/products` to `/add` to align with its purpose.